+++ outputs.tf
@@ -1,6 +1,6 @@
 output "s3_bucket_id" {
   description = "The name of the bucket."
-  value       = try(aws_s3_bucket_policy.this[0].id, aws_s3_bucket.this[0].id, "")
+  value       = try(aws_s3_bucket.this[0].id, aws_s3_directory_bucket.this[0].bucket, "")
 }
 
 output "s3_bucket_arn" {
@@ -35,7 +35,7 @@ output "s3_bucket_policy" {
 
 output "s3_bucket_region" {
   description = "The AWS region this bucket resides in."
-  value       = try(aws_s3_bucket.this[0].region, "")
+  value       = try(aws_s3_bucket.this[0].bucket_region, "")
 }
 
 output "s3_bucket_website_endpoint" {
@@ -46,4 +46,14 @@ output "s3_bucket_website_endpoint" {
 output "s3_bucket_website_domain" {
   description = "The domain of the website endpoint, if the bucket is configured with a website. If not, this will be an empty string. This is used to create Route 53 alias records."
   value       = try(aws_s3_bucket_website_configuration.this[0].website_domain, "")
+}
+
+output "s3_directory_bucket_name" {
+  description = "Name of the directory bucket."
+  value       = try(aws_s3_directory_bucket.this[0].bucket, null)
+}
+
+output "s3_directory_bucket_arn" {
+  description = "ARN of the directory bucket."
+  value       = try(aws_s3_directory_bucket.this[0].arn, null)
 }